Add async gens for multiple clans/players
I added these methods for:

 - `get_clan` -> `get_clans`
 - `get_current_war` -> `get_current_wars`
 - `get_player` -> `get_players`

I think that should be it. I also added examples in the docstrings as well. I figured `collections.Iterable` as a good enough description for people. And the `tags = iter(tags)` enforces that an iterable was passed before we even start iterating just as a check.