v3.9.1 (#267)
* Add Supercharger achievement

* Update pyproject.toml

* Update doc_requirements.txt

* Update conf.py

* Update changelog.rst

* Fix formatting of `**kwargs` in changelog entry

Corrected the markup surrounding `**kwargs` for consistency in the changelog. This ensures proper rendering in the documentation.

* Update version to 3.9.1

Bump the version from 3.9.0 to 3.9.1 to reflect recent changes. This ensures the software version aligns with the latest updates and modifications.

* Remove FullWar API extension from the codebase

The FullWar API extension and its associated documentation have been fully removed. This functionality was limited to a small group of users and is no longer maintained or required. Any references to it have been deleted to clean up the codebase.

* Remove `full_war_api` extension from documentation

The `full_war_api` extension has been removed due to lack of maintenance. Associated references in the documentation have also been deleted to reflect this change.

* Update minimum Python version to 3.9 across the project

Raise the required Python version from 3.7 to 3.9 in configuration files, build workflows, and documentation. As 3.7 and 3.8 are eol.

* Update changelog.rst

* Update enums.py